convert compiler to full dynamic typing, remove all type analysis.
This hurts the performance of the resulting output executable, but it results
in massive code savings in the compiler and the skeleton C code (something
like 400 fewer lines of code in total.) All variables and intermediate values
in the generated C code are now pyobj structs instead of primitive C types.
This required relatively few changes to the Python component, but the skeleton
C code needed a lot of changes to make that work.

Also do some cleanups to the skeleton code to fix compiler warnings.